Update semantics of strip_prefix argument.
strip_prefix was meant to preserve a given proxy-prefix while still
allowing the minwebhelpers library to do its filesystem magic. The key
is for the prefix to remain in place, but for the library to know the
location of the given resources in disk. This update instils this
desired functionality.